Implemented simple test for scene viewport model - prints the name of the clicked object to the command line on mouse down.

+++ src/tetzlaff/mvc/models/SceneViewportModel.java
@@ -2,5 +2,5 @@ package tetzlaff.mvc.models;
 
 public interface SceneViewportModel
 {
-	Object getObjectAtCoordinates(float x, float y);
+	Object getObjectAtCoordinates(double x, double y);
 }
